removed biometrics fields from UserDTO, added BiometricsDTO
