+++ src/main/java/org/training/food/tracker/dto/UserDTO.java
@@ -1,26 +1,18 @@
 package org.training.food.tracker.dto;
 
-import org.training.food.tracker.model.Lifestyle;
 import org.training.food.tracker.model.Role;
-import org.training.food.tracker.model.Sex;
 
 import java.math.BigDecimal;
 
 public class UserDTO {
 
-    private Long id;
     private String username;
     private String password;
     private String email;
     private String firstName;
     private String lastName;
     private Role role;
-    private BigDecimal age;
-    private Sex sex;
-    private BigDecimal weight;
-    private BigDecimal height;
-    private Lifestyle lifestyle;
-    private BigDecimal dailyNorm;
+    private BigDecimal dailyNormCalories;
 
     public static Builder builder() {
         return new Builder();
@@ -29,13 +21,8 @@ public class UserDTO {
     public static class Builder {
         private UserDTO userDTO;
 
-        public Builder() {
+        private Builder() {
             this.userDTO = new UserDTO();
-        }
-
-        public Builder id(Long id) {
-            userDTO.setId(id);
-            return this;
         }
 
         public Builder username(String username) {
@@ -68,47 +55,15 @@ public class UserDTO {
             return this;
         }
 
-        public Builder age(BigDecimal age) {
-            userDTO.setAge(age);
-            return this;
-        }
-
-        public Builder sex(Sex sex) {
-            userDTO.setSex(sex);
-            return this;
-        }
-
-        public Builder height(BigDecimal height) {
-            userDTO.setHeight(height);
-            return this;
-        }
-
-        public Builder weight(BigDecimal weight) {
-            userDTO.setWeight(weight);
-            return this;
-        }
-
-        public Builder lifestyle(Lifestyle lifestyle) {
-            userDTO.setLifestyle(lifestyle);
-            return this;
-        }
 
         public Builder dailyNorm(BigDecimal dailyNorm) {
-            userDTO.setDailyNorm(dailyNorm);
+            userDTO.setDailyNormCalories(dailyNorm);
             return this;
         }
 
         public UserDTO build() {
             return userDTO;
         }
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public String getUsername() {
@@ -159,51 +114,11 @@ public class UserDTO {
         this.role = role;
     }
 
-    public BigDecimal getAge() {
-        return age;
+    public BigDecimal getDailyNormCalories() {
+        return dailyNormCalories;
     }
 
-    public void setAge(BigDecimal age) {
-        this.age = age;
-    }
-
-    public Sex getSex() {
-        return sex;
-    }
-
-    public void setSex(Sex sex) {
-        this.sex = sex;
-    }
-
-    public BigDecimal getWeight() {
-        return weight;
-    }
-
-    public void setWeight(BigDecimal weight) {
-        this.weight = weight;
-    }
-
-    public BigDecimal getHeight() {
-        return height;
-    }
-
-    public void setHeight(BigDecimal height) {
-        this.height = height;
-    }
-
-    public Lifestyle getLifestyle() {
-        return lifestyle;
-    }
-
-    public void setLifestyle(Lifestyle lifestyle) {
-        this.lifestyle = lifestyle;
-    }
-
-    public BigDecimal getDailyNorm() {
-        return dailyNorm;
-    }
-
-    public void setDailyNorm(BigDecimal dailyNorm) {
-        this.dailyNorm = dailyNorm;
+    public void setDailyNormCalories(BigDecimal dailyNormCalories) {
+        this.dailyNormCalories = dailyNormCalories;
     }
 }
